+++ lib/builtin/output/output.dart
@@ -1,2 +1,4 @@
+export 'src/dock_status_item.dart';
+export 'src/extension.dart';
 export 'src/output_controller.dart';
 export 'src/output_view.dart';